requirements.txt
langchain==0.3.2
langgraph==0.2.5
crewai==0.51.1
jinja2==3.1.4
huggingface-hub==0.25.1
transformers==4.45.2
google-generativeai==0.8.3
elevenlabs==1.4.2
diffusers==0.30.3
pillow==10.4.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
celery==5.4.0
redis==5.0.8
apscheduler==3.10.4
pandas==2.2.3
streamlit==1.39.0
pytest==8.3.3
black==24.8.0
python-dotenv==1.0.1
asyncio-throttle==1.0.2
